Make the claw button toggle between open and close
This is just so that the driver doesn't have to hold down on the claw
close button to lift a cone since the cone kinda slips off of the claw
after a while. Allowing this to be toggled makes it easier for the
driver to both grab onto a claw and focus more on bringing it to a
mobile or stationary goal.

It's a good rule of thumb to keep functions small so I abstracted
everything into separate functions here and made them static to prevent
any possibility of a naming collision.