Add autonomous
The turning functions work, but straight/claw needs to be tested.

I used a ton of basic algebra and geometry to figure out how to do a
perfect turn relative to the robot's center of mass.

An autonomous debugging button was added to joystick 7L for testing the
autonomous without having to go through a control switch.

Part of the autonomous plan was implemented, but we think we might
wanna instead score our preload on the stationary goal so this may
change later.