Switch to C++
I'm better with C++ so we're gonna roll with it. The extra features like
classes and such help with writing less code to do more cool stuff.

The LCD's button state is now managed by an object to make it a bit
neater when writing code that uses it.

All of the LCD interface is managed by a single function, invoked inside
of its own little task started by init.cpp.

Those LCD loop functions taht control the interface now take a reference
to the ButtonState (to query buttons are being pressed) and they return
what they want the LoopState to be changed to. This helps cut down on
the amount of unnecessary global variables and instead migrating them to
local variables within lcdMain().

The enum AutonID was also changed to remove the AUTONID_COUNT constant
because it got in the way of auto.cpp's switch statement by emitting a
warning if there was no default case.

Moved global variable definitions to init.cpp.

Removed some unnecessary 'ul' things in auto.cpp.