Add lift PID controller
A PID (proportional, integral, derivative) controller uses IMEs to track
the lift's current position and keep it at a certain height. This runs
in its own task so it usually won't try to block anything. As of right
now, the PID isn't tuned yet, so it won't work just yet.

The LCD now has another interface setup, which is for controlling the
lift while the cortex is still plugged into the computer to get
debugging info from the PID controller.

Lift lock was removed to make way for PID.

All ports except for the lift are set to 0 for now until we finish
wiring everything up.

The include header <math.h> was added to main.hpp so we could use things
like the absolute value function.

Moved motor poll rate constant to main.hpp.