Add mobile goal lift (generalized) PID controller
A lot of the PID code was generalized to allow running a separate PID
controller in the same task. Also note that lift.cpp was renamed to
pid.cpp.

The struct pid::Module holds all the constants and variables involved in
doing the PID stuff. An internal array of these objects is iterated over
every time the PID updates itself to allow multiple controllers to exist
in a single task.